requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.0.1
beautifulsoup4==4.14.2
ccxt==4.4.69
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
executing==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.16
fastapi-cloud-cli==0.4.0
fastar==0.6.0
feedparser==6.0.12
filelock==3.19.1
fonttools==4.56.0
frozendict==2.4.7
frozenlist==1.5.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
iniconfig==2.3.0
IProgress==0.4
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.2.0
multitasking==0.0.12
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.5
packaging==24.2
pandas==2.3.3
parso==0.8.4
passlib==1.7.4
peewee==3.18.3
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.6.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==6.33.1
psutil==7.0.0
psycopg==3.3.0
psycopg-binary==3.3.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pybit==5.10.0
pycares==4.5.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.10
pydantic-settings==2.12.0
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
regex==2025.11.3
requests==2.32.3
rich==14.0.0
rich-toolkit==0.16.0
rignore==0.7.6
rsa==4.9.1
safetensors==0.6.2
Sastrawi==1.0.1
seaborn==0.13.2
sentry-sdk==2.45.0
setuptools==77.0.3
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.41.3
sympy==1.14.0
ta==0.11.0
tokenizers==0.22.1
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.8.14
uvicorn==0.32.1
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
yarl==1.18.3
yfinance==0.2.66
